Resize stateByID by 10% instead of a single element at a time
For a complex modpack, this reduces startup time by more than 99%, because Arrays.copyOf() is invoked only dozens of times instead of hundreds of thousands of times. Fixes #3284 and #3296, and possibly others.

Only patched forge 1.16.5 because that's all I'm running at the moment, but I expect the same fix can be applied elsewhere.