Revert "Merge pull request #4030 from ChimneySwift/generic-s3"
This reverts commit b0e56d3e5acc9419339f0adab755893f24098239, reversing
changes made to 20700c21b85afcd551e8aac1612f5e96fce7b7c8.

The reason for reverting this is that it breaks zoom tile generation for Amazon's s3 in an insidious way.
No errors are reported, and the zoom tile thread infinitely iterates over calling the s3 API,
throwing an error when s3 replies and then swallowing the error and immediately trying again.
This generates about 2-3MB/s of useless traffic to s3 and the immense amount of API calls made incurs a significant cost.